+++ CHANGELOG.md
@@ -7,6 +7,8 @@
     - Debug panel by pressing 'd'
     - Basic hallway drawing
     - Basic wooden doors, closed, functionally blocking you from walking down a halfway
+    - Light radius basics!
+    - Demo: killing enemies increases your lightradius
 
 ## 21-12-2021
     - Equipping weapon on right hand now impacts attack and damage calc